Introduce base clone controller and adjust PVC clone controller accordingly
We are introducing cloning from snapshot soon as a new controller,
as part of those efforts, it was identified to be beneficial to have this split.
