+++ cmd/cdi-controller/controller.go
@@ -213,7 +213,7 @@ func start(ctx context.Context, cfg *rest.Config) {
 		klog.Errorf("Unable to setup datavolume upload controller: %v", err)
 		os.Exit(1)
 	}
-	if _, err := dvc.NewCloneController(ctx, mgr, log,
+	if _, err := dvc.NewPvcCloneController(ctx, mgr, log,
 		clonerImage, importerImage, pullPolicy, getTokenPublicKey(), getTokenPrivateKey(), installerLabels); err != nil {
 		klog.Errorf("Unable to setup datavolume clone controller: %v", err)
 		os.Exit(1)
